These requirements are working for class c 24 block
"
Create IP addresses - take in a CIDR block 10.0.0.1/24 and add all IP addresses within that block to the data store with status “available” using regular expressions?
List IP addresses - return all IP addresses in the system with their current status
Acquire an IP - set the status of a certain IP to “acquired”
Release an IP - set the status of a certain IP to “available”
"

on a in memory database.

Use these to call the API:
http://localhost:8080/create?addressBlock=10.0.0.1/24
http://localhost:8080/list
http://localhost:8080/acquired?id=2
http://localhost:8080/available?id=2